move utils to libs and rename it to test
