+++ libs/design/components/CreateProject/ProjectForm/CreateProjectModal.tsx
@@ -1,5 +1,5 @@
 import { messages } from '@camp/messages';
-import { createTestAttr } from '@camp/utils';
+import { createTestAttr } from '@camp/test';
 import { closeModal, openModal } from '@mantine/modals';
 
 import { CreateProjectForm } from './CreateProjectForm';
