+++ libs/design/components/atoms/NavLink/ExitNavLink.tsx
@@ -1,6 +1,6 @@
 import { useAuth0 } from '@auth0/auth0-react';
 import { messages } from '@camp/messages';
-import { createTestAttr } from '@camp/utils';
+import { createTestAttr } from '@camp/test';
 import type { CSSObject, Sx } from '@mantine/core';
 import { NavLink as MantineNavLink } from '@mantine/core';
