+++ tsconfig.json
@@ -9,12 +9,12 @@
       "@camp/data-layer": ["./libs/data-layer/index.ts"],
       "@camp/design": ["./libs/design/index.ts"],
       "@camp/domain": ["./libs/domain/index.ts"],
+      "@camp/test": ["./libs/test/index.ts"],
       "@camp/config": ["./libs/config/index.ts"],
       "@camp/router": ["./libs/router/index.ts"],
       "@camp/messages": ["./app/messages.ts"],
       "@camp/hooks": ["./app/hooks/index.ts"],
-      "@camp/infra": ["./app/infra/index.ts"],
-      "@camp/utils": ["./app/utils/index.ts"]
+      "@camp/infra": ["./app/infra/index.ts"]
     }
   },
   "include": ["app", "@types", "libs"],
